pyproject.toml
[tool.poetry]
name = "photoshare"
version = "0.1.0"
description = ""
authors = ["Misha-HF <144258555+Misha-HF@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.13.1"
sqlalchemy = "^2.0.30"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
psycopg2 = "^2.9.9"
pydantic = "^2.7.1"
cloudinary = "^1.40.0"
pydantic-settings = "^2.2.1"
libgravatar = "^1.0.4"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.9"
bcrypt = "4.0.1"
qrcode = "^7.4.2"
pycryptodome = "^3.20.0"
pillow = "^10.3.0"
pytest = "^8.2.1"

httpx = "^0.27.0"
pytest-asyncio = "^0.23.7"
pytest-cov = "^5.0.0"



[tool.poetry.group.dev.dependencies]
sphinx = "^7.3.7"
sphinx-autodoc-typehints = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
